Fix input type for big number to long
Changed input method for 'big' from nextInt() to nextLong() to handle larger numbers.

+++ CSCI271_Assignment3_QuestionFour_LondynPepperdine.java
@@ -85,7 +85,7 @@ public class CSCI271_Assignment3_QuestionFour_LondynPepperdine
 
         //Read input number and digit from the user
         System.out.print("Enter number a big number: ");
-        long big = Scanner.nextInt();
+        long big = Scanner.nextLong();
         System.out.println("Enter one digit number: ");
         int small = Scanner.nextInt();
